Retire live-discovery and mid-run report code from progress_report
Deletes _progress_report_live.py and removes all live-run renderers
from progress_report.py. Retained helpers (_PhaseRound, _round_vendor_cost
chain, timing windows) are relocated into progress_report.py before deletion.
Tests for retired surfaces are removed; retained-surface coverage moves to
test_review_phase_detail.py.

Closes #6662
