+++ python/complexity-baseline.json
@@ -4146,24 +4146,6 @@
     "metric": 61
   },
   {
-    "file": "larch/report/_progress_report_live.py",
-    "code": "C901",
-    "qualified_symbol": "_latest_timing_ledger_activity_ts",
-    "metric": 11
-  },
-  {
-    "file": "larch/report/_progress_report_live.py",
-    "code": "C901",
-    "qualified_symbol": "_round_vendor_cost",
-    "metric": 14
-  },
-  {
-    "file": "larch/report/_progress_report_live.py",
-    "code": "PLR0912",
-    "qualified_symbol": "_round_vendor_cost",
-    "metric": 13
-  },
-  {
     "file": "larch/report/cleanup_implement_logs.py",
     "code": "C901",
     "qualified_symbol": "_upgrade_transcript",
@@ -4310,8 +4292,8 @@
   {
     "file": "larch/report/progress_report.py",
     "code": "C901",
-    "qualified_symbol": "_render_design_plan_review",
-    "metric": 13
+    "qualified_symbol": "_round_vendor_cost",
+    "metric": 14
   },
   {
     "file": "larch/report/progress_report.py",
@@ -4346,8 +4328,8 @@
   {
     "file": "larch/report/progress_report.py",
     "code": "PLR0912",
-    "qualified_symbol": "_render_design_plan_review",
-    "metric": 14
+    "qualified_symbol": "_round_vendor_cost",
+    "metric": 13
   },
   {
     "file": "larch/report/progress_report.py",
@@ -4360,12 +4342,6 @@
     "code": "PLR0913",
     "qualified_symbol": "_progress_vendor_rows",
     "metric": 7
-  },
-  {
-    "file": "larch/report/progress_report.py",
-    "code": "PLR0913",
-    "qualified_symbol": "_render_inflight_gantt",
-    "metric": 6
   },
   {
     "file": "larch/report/progress_report.py",
@@ -4384,12 +4360,6 @@
     "code": "PLR0913",
     "qualified_symbol": "render_phase_detail",
     "metric": 7
-  },
-  {
-    "file": "larch/report/progress_report.py",
-    "code": "PLR0915",
-    "qualified_symbol": "_render_design_plan_review",
-    "metric": 54
   },
   {
     "file": "larch/report/report_tokens_cost.py",
